Moved QuestAPI to `./src/control/public` / Created QuestDBShim.
