+++ src/control/FQLHooks.js
@@ -1,8 +1,8 @@
 import Socket           from './Socket.js';
-import QuestAPI         from './QuestAPI.js';
 import QuestDB          from './QuestDB.js';
 import Utils            from './Utils.js';
 import ViewManager      from './ViewManager.js';
+import QuestAPI         from './public/QuestAPI.js';
 import Quest            from '../model/Quest.js';
 import QuestFolder      from '../model/QuestFolder.js';
 import QuestCollection  from '../model/QuestCollection.js';
