+++ src/view/QuestTracker.js
@@ -1,5 +1,5 @@
 import RepositionableApplication from './RepositionableApplication.js';
-import QuestAPI                  from '../control/QuestAPI.js';
+import QuestAPI                  from '../control/public/QuestAPI.js';
 import QuestDB                   from '../control/QuestDB.js';
 import ViewManager               from '../control/ViewManager.js';
